resolving #696 comments & suggestions some API code refactoring
