+++ elenasrc3/elc/cliconst.h
@@ -13,7 +13,7 @@
 
 namespace elena_lang
 {
-   #define ELC_REVISION_NUMBER               0x005B
+   #define ELC_REVISION_NUMBER               0x005C
 
 #if defined _M_IX86 || _M_X64
