+++ elenasrc3/tools/asmc/asmconst.h
@@ -12,7 +12,7 @@
 namespace elena_lang
 {
 
-   #define ASM_REVISION_NUMBER               0x0005
+   #define ASM_REVISION_NUMBER               0x0006
 
    constexpr auto N_ARGUMENT1                = "__n_1";
    constexpr auto N_ARGUMENT2                = "__n_2";
