+++ examples60/db/sqlite/main.l
@@ -11,28 +11,28 @@ extension dbTestOp
 {
     printTable(tableName)
     {
-        console.printLine(tableName,":");
+        Console.printLine(tableName,":");
         var table := self.executeQuery("SELECT * FROM " + tableName);
         
         // Header
         table.fields().forEach::(fieldName)
         {
-            console.printPaddingRight(25, fieldName)
+            Console.printPaddingRight(25, fieldName)
         };
-        console.printLine();
-        console.printPaddingRightChar(25 * table.fieldCount(), $45, "-");
-        console.printLine();
+        Console.printLine();
+        Console.printPaddingRightChar(25 * table.fieldCount(), $45, "-");
+        Console.printLine();
         
         // Rows
         table.rows().forEach::(row)
         {
             table.fields().forEach::(fieldName)
             {
-                console.printPaddingRight(25, row[fieldName])
+                Console.printPaddingRight(25, row[fieldName])
             };
-            console.printLine()
+            Console.printLine()
         };
-        console.printLine();
+        Console.printLine();
     }
 }
 
@@ -50,28 +50,28 @@ public program()
   
         using (cnn)
         {
-            console.write("Database is being created");
+            Console.write("Database is being created");
             
             cnn.executeNonQuery("CREATE Table Movies (Name TEXT, Director TEXT, Year INTEGER)");
-            console.write(".");
+            Console.write(".");
                 
             cnn.executeNonQuery("INSERT INTO Movies (Name, Director, Year) VALUES ('The Dark Knight', 'Christopher Nolan', 2008)");
             cnn.executeNonQuery("INSERT INTO Movies (Name, Director, Year) VALUES ('Cloverfield', 'Matt Reeves', 2008)");
             cnn.executeNonQuery("INSERT INTO Movies (Name, Director, Year) VALUES ('Beverly Hills Chihuahua', 'Raja Gosnell', 2008)");
-            console.write(".");
+            Console.write(".");
                 
             cnn.executeNonQuery("CREATE TABLE [Users] " +
                                    "([ID] INTEGER  NOT NULL PRIMARY KEY AUTOINCREMENT, " +
                                    "[client] NVARCHAR(100)  NULL, " +
                                    "[date] DATE  NULL " +
                                    ")");
-            console.write(".");
+            Console.write(".");
                                    
             cnn.executeNonQuery("INSERT INTO Users (client, date) VALUES ('Alexandre', '2015-01-01')");
             cnn.executeNonQuery("INSERT INTO Users (client, date) VALUES ('Alex', '2015-01-01')");
-            console.write(".");
+            Console.write(".");
             
-            console.writeLine("Done")
+            Console.writeLine("Done")
         }                    
     };
 
@@ -79,9 +79,9 @@ public program()
 
     using (cnn)
     {
-        console.printLine("Number of table in DB: ",cnn.numberOfTable());
+        Console.printLine("Number of table in DB: ",cnn.numberOfTable());
     
-        console.writeLine("Tables:").writeLine();
+        Console.writeLine("Tables:").writeLine();
         
         cnn.tables().forEach::(tableName)
         {
@@ -89,5 +89,5 @@ public program()
         }
     };
 
-    console.readChar()
+    Console.readChar()
 }
